+++ src/main/scala/au/com/agl/arc/api/API.scala
@@ -244,6 +244,7 @@ object API {
     case class AmazonAccessKey(accessKeyID: String, secretAccessKey: String) extends Authentication
     case class AzureSharedKey(accountName: String, signature: String) extends Authentication
     case class AzureSharedAccessSignature(accountName: String, container: String, token: String) extends Authentication
+    case class AzureDataLakeStorageToken(clientID: String, refreshToken: String) extends Authentication
     case class GoogleCloudStorageKeyFile(projectID: String, keyFilePath: String) extends Authentication
   }