+++ src/main/scala/au/com/agl/arc/util/CloudUtils.scala
@@ -37,6 +37,17 @@ object CloudUtils {
           .field(s"fs.azure.sas.${container}.${accountName}.blob.core.windows.net", token)        
           .log()                    
       }
+      case Some(API.Authentication.AzureDataLakeStorageToken(clientID, refreshToken)) => {
+        hc.set("fs.adl.oauth2.access.token.provider.type", "RefreshToken")
+        hc.set("fs.adl.oauth2.client.id", clientID)
+        hc.set("fs.adl.oauth2.refresh.token", refreshToken)
+        logger.debug()
+          .message("hadoopConfiguration.set()")
+          .field("fs.adl.oauth2.access.token.provider.type", "RefreshToken")
+          .field("fs.adl.oauth2.client.id", clientID)
+          .field("fs.adl.oauth2.refresh.token", refreshToken)          
+          .log()                    
+      }
       case Some(API.Authentication.GoogleCloudStorageKeyFile(projectID, keyFilePath)) => {
         hc.set("google.cloud.auth.service.account.enable", "true")
         hc.set("fs.gs.project.id", projectID)